refactor(voice): split voice.rs into modular submodules
Split the 2074-line voice.rs into a voice/ module directory:
- adsr.rs: AdsrEnv modulation envelope
- modulation.rs: ModSources, mod_value_for, algo_param_slot_mods_for_line, modulated_line_params
- render.rs: render_voice and all rendering helpers
- mod.rs: Voice struct, LineEnvs, constants, tests

All 43 tests pass (2 pre-existing failures unrelated to this change).
